refactor: better name for new module and move it to new file
