+++ models/__init__.py
@@ -1,2 +1,3 @@
 from .agent_models import *  # noqa: F403
 from .latent_path_module import LatentPathModule  # noqa: F401
+from .place_cell_control_agent import PlaceCellControlModule  # noqa: F401
